_utilidades/requirements-CLQUIWD3011101.txt
Flask==3.0.0
reportlab==4.0.9
python-dotenv==1.0.0
waitress==3.0.2
pandas>=3.0.1
openpyxl>=3.1.5
